✨feat: rollup默认新增两种打包模式，优化配置方式。更接近原生，0学习成本；
