+++ .prettierrc.js
@@ -1,3 +1,3 @@
 module.exports = {
-  ...require('./src/configs/prettier-config'),
+  ...require('./src/prettier-config'),
 };
