+++ commitlint.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  extends: ['./src/configs/commitlint-config'],
+  extends: ['./src/commitlint-config'],
 };
