+++ example/app/Styled.tsx
