+++ example/lib/Foo/index.tsx
