+++ src/eslint-config/index.js
