+++ src/rollup-config/output.mjs
@@ -1,0 +1,32 @@
+/**
+ * @name 处理rollup导出文件
+ */
+const output = ({ outDir, buildType = 'lib' }) => {
+  const outputObj = {
+    app: [
+      {
+        file: `${outDir}/bundle.js`,
+        format: 'iife',
+        sourcemap: true,
+      },
+    ],
+    lib: [
+      // 打包为commonJS模块
+      {
+        dir: outDir,
+        format: 'cjs',
+        sourcemap: true,
+      },
+      // 打包为ES6模块
+      {
+        file: `${outDir}/index.esm.js`,
+        format: 'esm',
+        sourcemap: true,
+      },
+    ],
+  };
+
+  return outputObj[buildType];
+};
+
+export default output;
