+++ src/stylelint-config/index.js
